+++ src/converter.ts
@@ -16,14 +16,6 @@ export interface StateConverterOptionsWithContext
 export interface ConverterOptions<R, V> {
   convert(raw: R, options: StateConverterOptionsWithContext): V;
   render(value: V, options: StateConverterOptionsWithContext): R;
-  rawValidate?(
-    value: R,
-    options: StateConverterOptionsWithContext
-  ): boolean | Promise<boolean>;
-  validate?(
-    value: V,
-    options: StateConverterOptionsWithContext
-  ): boolean | Promise<boolean>;
   emptyRaw: R;
   emptyValue?: V;
   emptyImpossible?: boolean;
@@ -46,14 +38,12 @@ export interface IConverter<R, V> {
   preprocessRaw(raw: R, options: StateConverterOptionsWithContext): R;
 }
 
-export class ConvertError {}
-
 export class ConversionValue<V> {
   constructor(public value: V) {}
 }
 
 export class ConversionError {
-  constructor(public type: string) {}
+  constructor(public type: string = "default") {}
 }
 
 export type ConversionResponse<V> = ConversionError | ConversionValue<V>;
@@ -96,30 +86,12 @@ export class Converter<R, V> implements IConverter<R, V> {
     raw: R,
     options: StateConverterOptionsWithContext
   ): Promise<ConversionResponse<V>> {
-    if (this.definition.rawValidate) {
-      const rawValidationSuccess = await this.definition.rawValidate(
-        raw,
-        options
-      );
-      if (!rawValidationSuccess) {
-        return CONVERSION_ERROR;
-      }
-    }
     try {
       const value = this.definition.convert(raw, options);
-      if (this.definition.validate) {
-        const rawValidationSuccess = await this.definition.validate(
-          value,
-          options
-        );
-        if (!rawValidationSuccess) {
-          return CONVERSION_ERROR;
-        }
-      }
       return new ConversionValue<V>(value);
     } catch (e) {
-      if (e instanceof ConvertError) {
-        return CONVERSION_ERROR;
+      if (e instanceof ConversionError) {
+        return e;
       }
       throw e;
     }