+++ test/converter.test.ts
@@ -1,12 +1,7 @@
 import { configure } from "mobx";
 import { types } from "mobx-state-tree";
 import { Field, Form, converters, FieldAccessor } from "../src";
-import {
-  CONVERSION_ERROR,
-  ConversionValue,
-  Converter,
-  ConvertError
-} from "../src/converter";
+import { ConversionValue, Converter, ConversionError } from "../src/converter";
 
 configure({ enforceActions: "observed" });
 
@@ -52,8 +47,12 @@ test("converter to integer", async () => {
   const converter = new Converter<string, number>({
     emptyRaw: "",
     emptyImpossible: true,
-    rawValidate: raw => /^\d+$/.test(raw),
-    convert: raw => parseInt(raw, 10),
+    convert: raw => {
+      if (!/^\d+$/.test(raw)) {
+        throw new ConversionError();
+      }
+      return parseInt(raw, 10);
+    },
     render: value => value.toString()
   });
 
@@ -62,46 +61,7 @@ test("converter to integer", async () => {
   expect((result as ConversionValue<number>).value).toEqual(3);
 
   const result2 = await converter.convert("not a number", options);
-  expect(result2).toEqual(CONVERSION_ERROR);
-});
-
-test("converter with validate", async () => {
-  const converter = new Converter<string, number>({
-    emptyRaw: "",
-    emptyImpossible: true,
-    convert: raw => parseInt(raw, 10),
-    render: value => value.toString(),
-    validate: value => value <= 10
-  });
-
-  const result = await converter.convert("3", options);
-  expect(result).toBeInstanceOf(ConversionValue);
-  expect((result as ConversionValue<number>).value).toEqual(3);
-
-  const result2 = await converter.convert("100", options);
-  expect(result2).toEqual(CONVERSION_ERROR);
-});
-
-test("converter with async validate", async () => {
-  const done: any[] = [];
-
-  const converter = new Converter<string, string>({
-    emptyRaw: "",
-    emptyValue: "",
-    convert: raw => raw,
-    validate: async value => {
-      await new Promise(resolve => {
-        done.push(resolve);
-      });
-      return true;
-    },
-    render: value => value
-  });
-
-  const result = converter.convert("foo", options);
-  done[0]();
-  const v = await result;
-  expect((v as ConversionValue<string>).value).toEqual("foo");
+  expect(result2).toBeInstanceOf(ConversionError);
 });
 
 test("converter maybeNull with converter options", async () => {
@@ -129,21 +89,21 @@ test("converter maybeNull with converter options", async () => {
   expect(field.value).toEqual("36365.20");
 });
 
-test("convert can throw ConvertError", async () => {
+test("convert can throw ConverterError", async () => {
   const converter = new Converter<string, string>({
     emptyRaw: "",
     emptyValue: "",
     convert: raw => {
-      throw new ConvertError();
+      throw new ConversionError();
     },
     render: value => value
   });
 
   const result = await converter.convert("foo", options);
-  expect(result).toEqual(CONVERSION_ERROR);
+  expect(result).toBeInstanceOf(ConversionError);
 });
 
-test("non-ConvertError bubbles up", async () => {
+test("non-ConverterError bubbles up", async () => {
   const converter = new Converter<string, string>({
     emptyRaw: "",
     emptyValue: "",